Move the prioritiesEdit state up
To be able to reset its values when form is submitted
