+++ todo-list-app/src/components/Forms/EditTaskForm/EditTaskForm.js
@@ -1,46 +1,15 @@
 import './EditTaskForm.css';
 import { GrClose, GrTrash } from 'react-icons/gr';
 import { FaFlag } from 'react-icons/fa';
-import { 
-    BsFillSquareFill, 
-    BsFillCircleFill, 
-    BsFillTriangleFill, 
-    BsFillStarFill 
-} from 'react-icons/bs';
 import { useState, useEffect } from 'react';
-
-const prioritiesListEdit = [
-    {
-      value: 1,
-      icon: <BsFillStarFill color='red' size='1.2rem'/>,
-      id: 'first-priority-edit',
-      status: '',
-    },
-    {
-      value: 2,
-      icon: <BsFillSquareFill color='orange' size='1.2rem'/>,
-      id: 'second-priority-edit',
-      status: '',
-    },
-    {
-      value: 3,
-      icon: <BsFillTriangleFill color='blue' size='1.2rem'/>,
-      id: 'third-priority-edit',
-      status: '',
-    },
-    {
-      value: 4,
-      icon: <BsFillCircleFill color='#efefef' size='1.2rem'/>,
-      id: 'fourth-priority-edit',
-      status: 'selected',
-    },
-];
 
 function EditTaskForm({ 
     toggleEditForm, 
     editTodo, 
     setEditTodo, 
-    handleOnChangeEditForm 
+    handleOnChangeEditForm,
+    prioritiesEdit,
+    setPrioritiesEdit 
 }) {
     function handleOverlayClicked(e) {
         if (e.target.className === 'edit-task-overlay') {
@@ -48,8 +17,6 @@ function EditTaskForm({
         };
     };
     
-    const [prioritiesEdit, setPrioritiesEdit] = useState(prioritiesListEdit);
-
     function handlePrioritySelected(e) {
         const targetID = e.target.dataset.id;
         const newPrioritiesEdit = prioritiesEdit.map(priority => {
